+++ stream.h
@@ -2,8 +2,8 @@ typedef struct {
     unsigned char buffer;
     unsigned int buffer_size;
 
-    unsigned int total_length;
-    unsigned int alloced_space;
+    size_t total_length;
+    size_t alloced_space;
     unsigned char* data;
 } BitStream;
 
@@ -11,5 +11,5 @@ BitStream* create_stream();
 void write_bit(BitStream* stream, unsigned int bit);
 void write_byte(BitStream* stream, unsigned char byte);
 void flush(BitStream* stream);
-unsigned char* get_stream_data(BitStream* stream, unsigned int* bytes);
+unsigned char* get_stream_data(BitStream* stream, size_t* bytes);
